rewrite events to small modules
